requirements.txt
pytest-mozwebqa==0.10
requests==0.12.1
